build: rocm_clang_toolchain.cmake policy flag
Add code to create a toolchain policy flag file
so build script can check for it and proceed or
ignore the build for the distro/version.

This will avoid the CI build scripts to treat
'No rule to make target' as an error.

Code changes related to the following:
  * rocm_clang_toolchain.cmake

Note:
  * The CI build script 'build_rocm-bandwidth-test.sh' relies on this

Signed-off-by: Oliveira, Daniel <daniel.oliveira@amd.com>